app/requirements.txt
speechrecognition>=3.10.0
pyaudio>=0.2.11
pyserial>=3.5
numpy>=1.21.0